+++ src/grandorgue/GOBitmapCache.cpp
@@ -11,10 +11,10 @@
 #include <wx/intl.h>
 #include <wx/mstream.h>
 
+#include "files/GOOpenedFile.h"
 #include "loader/GOLoaderFilename.h"
 
 #include "GOBuffer.h"
-#include "GOFile.h"
 #include "GOLog.h"
 #include "GOOrganController.h"
 #include "Images.h"
@@ -172,9 +172,10 @@ bool GOBitmapCache::loadFile(wxImage &img, const wxString &filename) {
 
   try {
     GOLoaderFilename name;
-    name.Assign(m_OrganController->GetFileStore(), filename);
+    name.Assign(filename);
 
-    std::unique_ptr<GOFile> file = name.Open();
+    std::unique_ptr<GOOpenedFile> file
+      = name.Open(m_OrganController->GetFileStore());
     GOBuffer<char> data;
 
     result = file->ReadContent(data);